task1 updated
fixed names
